Split form into multiple components
